+++ src/widget.rs
@@ -14,6 +14,7 @@ pub mod hsv_selector;
 pub mod manipulate;
 pub mod manipulate_tool;
 pub mod move_camera;
+pub mod move_frame;
 pub mod move_tool;
 pub mod number_box;
 pub mod pixel_canvas;